filters.py from last year
ok i think i understand the architecture now?
danger zone does odometry in the autonav_filters node, and it's structured
as a main script that runs the show and one or more distinct filters (dead
reckoning and particle filter). also there's c++ for some reason and it
seems less organized than and unrelated to the python stuff. somehow. this
autonav_odometry could be similar but python only and with just a particle
filter to start